+++ src/common/default.ts
@@ -1,0 +1,11 @@
+export const oneHour = 1000 * 60 * 60;
+export const oneDay = oneHour * 24;
+export const oneMonth = oneDay * 30;
+
+export const subdomain = 'vscode';
+
+export const vscodeTemplate = {
+    id: 'dtmi:zyprqt7wqcw:n5mrwvbjpuc',
+    solutionModelId: 'dtmi:zyprqt7wqcw:xrgufh3yfz',
+    capabilityModelId: 'dtmi:vscode:VSCode1t3;2'
+};