+++ src/extension.ts
@@ -1,37 +1,23 @@
-// The module 'vscode' contains the VS Code extensibility API
-// Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
 import { connectToCentral } from './service/central';
 import { login } from './service/login';
 
-// this method is called when your extension is activated
-// your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
-
-	// Use the console to output diagnostic information (console.log) and errors (console.error)
-	// This line of code will only be executed once when your extension is activated
-	console.log('Congratulations, your extension "iotc" is now active!');
-
-	// The command has been defined in the package.json file
-	// Now provide the implementation of the command with registerCommand
-	// The commandId parameter must match the command field in package.json
 	context.subscriptions.push(vscode.commands.registerCommand(
-		'iotc.helloWorld',
+		'iotc.stop',
 		() => {
-			// The code you place here will be executed every time your command is executed
-			// Display a message box to the user
-			vscode.window.showInformationMessage('Hello World from iotc!');
+			vscode.window.showInformationMessage('Stopped');
 		}
 	));
 
 	context.subscriptions.push(vscode.commands.registerCommand(
-		'iotc.login',
+		'iotc.start',
 		async () => {
 			const result = await vscode.window.showInformationMessage('Do you want to login to Azure and connect IoTCentral?', 'Yes', 'No');
 			if (result === 'Yes') {
 				await login();
 				await connectToCentral();
-				vscode.window.showInformationMessage('Logged into Azure!');
+				vscode.window.showInformationMessage('Connected to IoTCentral https://vscode.azureiotcentral.com');
 			}
 			else {
 				console.log('log no');
@@ -48,5 +34,4 @@ export function activate(context: vscode.ExtensionContext) {
 	// connectToCentral();
 }
 
-// this method is called when your extension is deactivated
 export function deactivate() { }