+++ CHANGELOG.md
@@ -7,3 +7,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [0.1.0]
 
 - Initial release.
+
+## [0.2.0]
+- Added login message
+- Provision device instance